pyproject.toml
[project]
name = "conversions-api-app"
version = "0.1.0"
description = "Databricks App that gives Meta's advertiser customers a guided, UI-driven way to send server-side conversion events from their Databricks Lakehouse to Meta's Conversions API (CAPI)."
readme = "README.md"
license = { text = "Databricks License" }
requires-python = ">=3.11"
keywords = [
    "databricks",
    "databricks-labs",
    "databricks-apps",
    "meta",
    "facebook",
    "conversions-api",
    "capi",
    "marketing",
    "ad-tech",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Office/Business :: Financial :: Accounting",
]
authors = [
    { name = "Amelia Chu", email = "amelia.chu@databricks.com" },
    { name = "Jason Pohl", email = "jason@databricks.com" },
]
maintainers = [
    { name = "Amelia Chu", email = "amelia.chu@databricks.com" },
    { name = "Jason Pohl", email = "jason@databricks.com" },
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "databricks-sdk>=0.30.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.6",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/databrickslabs/conversions-api-app"
Repository = "https://github.com/databrickslabs/conversions-api-app"
Issues = "https://github.com/databrickslabs/conversions-api-app/issues"
Changelog = "https://github.com/databrickslabs/conversions-api-app/blob/main/CHANGELOG.md"

[tool.uv]

[[tool.uv.index]]
name = "databricks-proxy"
url = "https://pypi-proxy.dev.databricks.com/simple/"
default = true

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
# Put the app subfolder on sys.path so tests can `from server.X import ...`
pythonpath = ["app"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]

[tool.coverage.run]
source = ["app/server"]
omit = [
    "tests/*",
    "**/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true
skip_covered = false
